requirements.txt
streamlit
pandas
numpy
joblib
pickle-mixin
plotly
fpdf2
Pillow